+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/settings/AvatarCutterViewModel.kt
@@ -1,26 +1,29 @@
 package de.connect2x.trixnity.messenger.viewmodel.settings
 
 import com.arkivanov.essenty.backhandler.BackCallback
+import de.connect2x.trixnity.messenger.i18n.I18n
 import de.connect2x.trixnity.messenger.util.FileDescriptor
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
-import de.connect2x.trixnity.messenger.viewmodel.i18n
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import net.folivo.trixnity.client.media
+import net.folivo.trixnity.core.model.RoomId
+import net.folivo.trixnity.core.model.events.m.room.AvatarEventContent
+import org.koin.core.component.get
 
 
-private val log = KotlinLogging.logger { }
-
+private val log = KotlinLogging.logger {}
 
 interface AvatarCutterViewModelFactory {
     fun create(
         viewModelContext: MatrixClientViewModelContext,
         file: FileDescriptor,
+        roomId: RoomId? = null,
         onClose: () -> Unit,
     ): AvatarCutterViewModel {
-        return AvatarCutterViewModelImpl(viewModelContext, onClose = onClose, file = file)
+        return AvatarCutterViewModelImpl(viewModelContext, onClose = onClose, file = file, roomId = roomId)
     }
 
     companion object : AvatarCutterViewModelFactory
@@ -30,6 +33,7 @@ interface AvatarCutterViewModel {
     val upload: StateFlow<Boolean>
     val error: StateFlow<String?>
     val file: FileDescriptor
+    val avatarCutterHeading: String
     fun cancel()
     fun accept()
 }
@@ -38,10 +42,16 @@ open class AvatarCutterViewModelImpl(
     viewModelContext: MatrixClientViewModelContext,
     override val file: FileDescriptor,
     private val onClose: () -> Unit,
+    private val roomId: RoomId?
 ) : MatrixClientViewModelContext by viewModelContext, AvatarCutterViewModel {
 
+    private val i18n = get<I18n>()
     override val upload = MutableStateFlow(false)
     override val error = MutableStateFlow<String?>(null)
+
+    override val avatarCutterHeading =
+        if (roomId == null) i18n.yourNewProfileAvatar()
+        else i18n.yourNewRoomAvatar()
 
     private val backCallback = BackCallback {
         cancel()
@@ -58,31 +68,63 @@ open class AvatarCutterViewModelImpl(
     override fun accept() {
         coroutineScope.launch {
             upload.value = true
-            matrixClient.media.prepareUploadThumbnail(
+            val preparedUpload = matrixClient.media.prepareUploadThumbnail(
                 file.content,
                 file.mimeType,
-            )?.let { (cache, _) ->
-                matrixClient.media.uploadMedia(cache).fold(
-                    onSuccess = { url ->
-                        matrixClient.setAvatarUrl(url).fold(
-                            onSuccess = {
-                                upload.value = false
-                                onClose()
-                            },
-                            onFailure = {
-                                log.error(it) { "Cannot set user avatar." }
-                                upload.value = false
-                                error.value = i18n.profileAvatarError()
-                            }
-                        )
-                    },
-                    onFailure = {
-                        log.error(it) { "Cannot upload avatar image." }
-                        upload.value = false
-                        error.value = i18n.profileAvatarError()
-                    }
-                )
+            )
+
+            if (preparedUpload == null) {
+                log.error { "Failed to prepare upload thumbnail." }
+                upload.value = false
+                error.value = i18n.profileAvatarError()
+                return@launch
             }
+
+            val (cache, _) = preparedUpload
+
+            matrixClient.media.uploadMedia(cache).fold(
+                onSuccess = { url ->
+                    if (roomId == null) setUserAvatar(url)
+                    else setRoomAvatar(url, roomId)
+                },
+                onFailure = {
+                    log.error(it) { "Cannot upload avatar image." }
+                    upload.value = false
+                    error.value = i18n.profileAvatarError()
+                }
+            )
         }
     }
+
+    private suspend fun setUserAvatar(url: String) {
+        matrixClient.setAvatarUrl(url).fold(
+            onSuccess = {
+                upload.value = false
+                onClose()
+            },
+            onFailure = {
+                log.error(it) { "Cannot set user avatar." }
+                upload.value = false
+                error.value = i18n.profileAvatarError()
+            }
+        )
+    }
+
+    private suspend fun setRoomAvatar(url: String, roomId: RoomId) {
+        matrixClient.api.room.sendStateEvent(
+            roomId,
+            eventContent = AvatarEventContent(url = url),
+        ).fold(
+            onSuccess = {
+                upload.value = false
+                onClose()
+            },
+            onFailure = {
+                log.error(it) { "Cannot set room avatar." }
+                upload.value = false
+                error.value = i18n.profileAvatarError()
+            }
+        )
+    }
 }
+