+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Editable room avatar
+
 ### Changed
+- upgrade trixnity version
 
 ### Deprecated